+++ ash/src/extensions/ext/debug_report.rs
@@ -16,7 +16,7 @@ impl crate::ext::debug_report::Instance {
         (self.fp.destroy_debug_report_callback_ext)(
             self.handle,
             debug,
-            allocation_callbacks.as_raw_ptr(),
+            allocation_callbacks.to_raw_ptr(),
         );
     }
 
@@ -31,7 +31,7 @@ impl crate::ext::debug_report::Instance {
         (self.fp.create_debug_report_callback_ext)(
             self.handle,
             create_info,
-            allocation_callbacks.as_raw_ptr(),
+            allocation_callbacks.to_raw_ptr(),
             debug_cb.as_mut_ptr(),
         )
         .assume_init_on_success(debug_cb)