+++ ash/src/extensions/khr/android_surface.rs
@@ -17,7 +17,7 @@ impl crate::khr::android_surface::Instance {
         (self.fp.create_android_surface_khr)(
             self.handle,
             create_info,
-            allocation_callbacks.as_raw_ptr(),
+            allocation_callbacks.to_raw_ptr(),
             surface.as_mut_ptr(),
         )
         .assume_init_on_success(surface)